requirements/requirements.txt
eyeD3>=0.9.5
nicfit.py>=0.8.6
PyGObject
